requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via uvicorn
exrex==0.10.5
    # via pydantic-factories
faker==12.1.0
    # via pydantic-factories
h11==0.13.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
jsonpatch==1.32
    # via -r requirements.in
jsonpointer==2.2
    # via jsonpatch
loguru==0.6.0
    # via -r requirements.in
openapi-schema-pydantic==1.2.2
    # via starlite
orjson==3.6.6
    # via starlite
passlib==1.7.4
    # via -r requirements.in
pydantic==1.9.0
    # via
    #   openapi-schema-pydantic
    #   pydantic-factories
    #   starlite
pydantic-factories==1.2.3
    # via starlite
python-dateutil==2.8.2
    # via faker
python-multipart==0.0.5
    # via starlite
pyyaml==6.0
    # via starlite
requests==2.27.1
    # via starlite
ruamel-yaml==0.17.20
    # via -r requirements.in
ruamel-yaml-clib==0.2.6
    # via ruamel.yaml
six==1.16.0
    # via
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via anyio
starlette==0.18.0
    # via starlite
starlite==1.0.5
    # via -r requirements.in
toml==0.10.2
    # via -r requirements.in
typing-extensions==4.0.1
    # via
    #   pydantic
    #   pydantic-factories
    #   starlite
urllib3==1.26.8
    # via requests
uvicorn==0.17.4
    # via -r requirements.in
